Port away from KLocale (which was used for use12Clock()).
Amazingly, this fixes a bug with a messagebox showing "Yes/No" instead
of "Merge/Add" when launching korganizer with a .ics file as argument.

Here's why:
KLocale::global() creates a QTranslator instance, which sends a LanguageChange
event. QDialogButtonBox (already shown at that point) reacts by retranslating
the buttons, ignoring the custom text we set on the buttons. That's a Qt bug,
I'll look into fixing that as well.

BUG: 357556
FIXED-IN: Applications/16.04