Fix color inconsistency for unset categories.
Summary:
KOrganizer's preference dialog states that the color for unset
categories is only used on the categories-color-only setting. However,
this is currently not true for Month View, where in each of the mixed
settings (category-border, resource-fill, or vice versa) the
unset-category color is also used. With this patch, Month View considers
the unset-category color only in the category-only setting, just like
Agenda View.

BUG: 382185
FIXED-IN: 5.9.3

Test Plan:
By testing all four combinations for coloring frame and fill
a month item, I confirmed Bug 382185: the color for items with no
categories is used in CategoryOnly as well as in mixed settings (i.e.,
all settings but ResourceOnly). After applying this patch, I tested them
again and confirmed that the unset-category color is only used in
CategoryOnly, while in mixed settings, the resource color is used instead,
making it consistent to Agenda View and the settings dialog.

Reviewers: #kde_pim, dvratil

Reviewed By: #kde_pim, dvratil

Subscribers: dvratil, kde-pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D16088