Eventviews: properly fall back to color-generator if ColorAttribute is absent
Summary:
This was introduced by 1ec2e08 which only called the color-generator
code when there *is* a ColorAttribute on the collection (with an
invalid color, which seems unlikely), rather than when the attribute
isn't there.

I thought this was the bug that made all new resources red (with caldav)
but actually that seems to come from the resource itself. So now I'm not
100% sure about this commit anymore, although it still seems to make
sense to me -- for other resources, possibly.

Test Plan:
Disabling the code that reads from the ColorAttribute
and removing all "Resources Colors" entries from ~/.config/eventviewsrc,
the colors are properly generated.

Reviewers: ochurlaud, dkurz, dvratil

Reviewed By: ochurlaud, dkurz

Subscribers: #kde_pim

Differential Revision: https://phabricator.kde.org/D16508