agenda/agenda.cpp - grid colors when not using system settings
this addresses https://bugs.kde.org/show_bug.cgi?id=411608
which refers to a much older KOrganizer version 5.9.2.

Newer KOrganizer handles grid color (computed from the
windowtext) just fine when using system settings; but,
not when using customer background colors.  With this
patch we fix that.

The bug report requested a custom configuration for the
grid color.  The UX team doesn't want that and requested
to remove the ability for custom colors completely.
That's a different thread and a different fight.

BUG: 411608
FIXED-IN 5.12.3

Differential Revision: https://phabricator.kde.org/D411608