Display all-day to-dos on their due date in the agenda view
The agenda view displays non-recurring all-day to-dos that are not overdue
on the day before their due date.  Recurring to-dos and non-all-day to-dos
are displayed correctly, on their due dates.

`displayIncidence()` deliberately displays to-dos due at midnight in the
event view at the end of the previous day, but that should not be done for
all-day to-dos.  This patch adds an all-day check.

The patch opportunistically removes two `setTime()` calls that duplicate
work done in initializers, and tidies a third.

BUG: 338302

BUG: 417982