Display recurring to-dos on their due date/time in the agenda view
The agenda view displays non-recurring to-dos on their due date, and at
their due time-of-day if they are not all-day todos.  However, it displays
all-day recurring to-dos on their *start* date, and other recurring to-dos on
their *start* date but at their *due* time-of-day.

This patch displays all to-dos on their due date.