Preserve duration when dragging to-dos in the Agenda view
When a to-do in the time-of-day part of the agenda view is dragged, its
start time is reset to 30 minutes earlier than the due time, because the
agenda view bases its calculations on the size of the graphical item.
This patch shifts the start and the next recurrence, if present, to match
the change in the due date.