agendaviews: Display recurrence w/o main incidence
Problem: Currently the agenda view hides recurrences which no longer
have a main incidence attached to it. Roundcube just displays these
recurrences, which seems to preferable compared to simply hiding an
event.

This patch forces the agenda view to still display a "lost" recurrence.

Note about the patch: Removed the `mViewCalendar->isValid(incidence)`
check in the if-condition since this is already checked in the first
line of the function and simplifies code.