Adjust layout and resizing of the timeline view
If KOrganizer shows the agenda view when it launches, and then is
switched to the timeline view, the timeline gives too much space to
the calendar list and too little to the time grid.  Also, when the
main window is resized, the calendar list is not resized.

This patch sets the initial panel sizes and resize behavour, which
makes `splitterMoved()` redundant.