Display multi-day holidays as one wide block in the Month View
Display multi-day holidays, such as Lunar New Year in the Hong Kong
holiday list, just like a multi-day event:  a block that spans days,
instead of separate blocks on each day.

As a side-effect, Valentine's Day does not appear in the Month View
when it happens to fall on a non-workday.  The Month View will only
show true non-workday holidays.

Users of multiple regional calendars may see multiple holidays on a
day, instead of one block with multiple holiday names in it.

BUG: 187116