Agenda: fix timezone being lost when moving event by drag
The code was using QDateTime::toTimeSpec() in attempt to restore the
timezone from the original dtStart/dtEnd. However, QDateTime::toTimeSpec()
automatically converts Qt::TimeZone timespec to Qt::LocalTime, so the
information about timezone was lost and the time was treated as local.

The solution is to restore the timezone manually when the original
timespec was Qt::TimeZone.

CCBUG: 448703
FIXED-IN: 5.24.1