ListView: fix list view choking on large calendars
The list view was generating tooltip string for each incidence
upon insertion (and multiple times - once for each column) without
providing current date hint, which caused the tooltip routine to
loop over all recurrences. In large calendars with many recurrent
events this could freeze the app for minutes.

The new code overrides the QTreeViewItem::data() and generates the
tooltip string on demand, which also saves a substantial amount of
memory.