Apply current view filter on the Todo model as well
Previously the Todo model was derived from ETMCalendar, which internally
had a CalFilterProxyModel to filter out incidences based on a CalFilter.
Nowadays we are building right on top of an ETM, so we need to apply the
filter manually.

It's a little hacky solution, but it is achieved by taking the current
filter from one of the active calendars and manually filtering each
Todo through it in the TodoViewSortFilterProxyModel.

CCBUG: 484040
FIXED-IN: 24.02.3