Don't overwrite valid date in eventDurationHint()
The view's hint function is called whenever a new event/todo/journal
is being created from KOrganizer. When the new incidence is being
created via context menu in the side-pane calendar preview, the selected
date might be different from the date currently selected in the main
view. In such case, the view should only provide a hint (e.g. about
duration or allday-ness of the new event) but shouldn't replace the
valid dates.

BUG: 483823
FIXED-IN: 24.02.2