todoview.cpp - add note, tooltip. whatsthis about "quick add" to-dos
"Quick-add" to-dos are open-ended.

You would think that we should add default reminders
(if one is configured) for a quick to-dos; except that
quick to-dos don't set a start datetime or due datetime;
and without a start datetime you can't have a reminder.

IOW: quick add todos are open-ended and therefore
we have no time reference for a reminder.